add squad-local-bisect script
This script will create a bisection script that should be ran from the
kernel directory.
$ ./squad-local-bisect --group lkft \
   --project linux-next-master \
   --build next-20230110 \
   --fail qemu-arm64/gcc-12-lkftconfig/boot/gcc-12-lkftconfig

Finding a failed test/build can be done with the script squad-compare-builds.
$ ./squad-compare-builds --group lkft \
   --project linux-next-master \
   --build next-20230109,next-20230110 --table

The squad-local-bisect script will create a sh script like this:
bisect-badsha-next-20230110-qemu-arm64-gcc-12-lkftconfig-boot-gcc-12-lkftconfig.sh

This script should run from /path/to/linux/tree/ and the GOOD_SHA should
be passed in as a env variable for now.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>