squad-compare-builds: compare more than too builds
To compare two or more builds with squad-compare-builds will be easier
with the flag '--gpb <group> <project> <build>" and pass the in two or
more time to see the comparison in a table format.

Example of how to run the script:

./squad-compare-builds --gpb lkft linux-stable-rc-linux-6.1.y latest \
    --gpb lkft linux-mainline-master latest \
    --gpb lkft linux-mainline-master v6.2

This patch dropped difflib also, it only supports the table format now.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>