Add support for creating LTP skipfile reproducers
Add the squad-create-skipfile-reproducers script to create TuxRun
reproducers for testing the LTP skipfile entries.
By providing the group and optionally the branches, devices and build
names, the script will create reproducers for all combinations of these
parameters. If any of these parameters are not provided, the default
behaviour is to run will all supported values for the given parameter.
Extra parameters include:
  * a --allow-unfinished flag to allow fetching of
    reproducers where the build has not been marked as finished.
  * A --suite-name parameter to set which LTP suite reproducer should be
    fetched when creating reproducers.
  * A --debug flag to add extra debug prints when the script runs.

Signed-off-by: Katie Worton <katie.worton@linaro.org>