read-skipfile-results: Create script for reading ltp skipfile test results
Create script for reading skipfile results as part of the skipfile
pipeline.

Tux plan files generated by squad-create-skipfile-reproducers are
submitted by the Gitlab CI pipeline, which logs the SQUAD build names
for the results into a text file. The read-skipfile-results script then
loops through the list of SQUAD builds and reads the results of the
skipfile testing (counting the non-hang tests in a dataframe). For each
test, the number of non-hang runs is compared to the total number of
runs, to determine which skipfile entries are safe to remove. Then, the
script will update the skipfile in the test-definitions GitHub repo and
optionally push these updates as PRs to the test-definitions repo. The
results dataframes are also stored to csv files in test-definitions for
inspection (not tracked by git).

Signed-off-by: Katie Worton <katie.worton@linaro.org>