.pre-commit-config.yaml: Add pre-commit config
Add a pre-commit config to the repo to make it easier to run the
relevant formatters and linters which are run in the CI pipeline for
this repo.

This will make it easier to spot linter and formatter errors as the
tooling which runs in the CI pipeline can be run locally in one step.

Signed-off-by: Katie Worton <katie.worton@linaro.org>